% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datHBA.donor5}
\alias{datHBA.donor5}
\title{Micro RNA Expression Values for the Human Brain Atlas.}
\format{A data frame with 14594631 rows and 4 variables:
\describe{
 \item{donorID}{brain ID from the Allen Institute}
 \item{gene}{gene the above probe corresponds to; gene to probe is many to one}
 \item{brain_structure}{brain structure acronym}
 \item{value}{normalized expression value}
}}
\source{
\url{"http://human.brain-map.org/static/download"}
}
\usage{
datHBA.donor5
}
\description{
A dataset containing the normalized expression values for a single brain in the human brain atlas.
}
\keyword{datasets}
