% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datHBA}
\alias{datHBA}
\title{Micro RNA Expression Values for the Human Brain Atlas.}
\format{A data frame with 32160624 rows and 5 variables:
\describe{
 \item{donorID}{brain ID from the Allen Institute}
 \item{probe_name}{mRNA probe used for this observation}
 \item{gene}{gene the above probe corresponds to; gene to probe is many to one}
 \item{brain_structure}{brain structure acronym}
 \item{value}{normalized expression value}
}}
\source{
\url{"http://human.brain-map.org/static/download"}
}
\usage{
datHBA
}
\description{
A dataset containing the normalized expression values for the human brain atlas.
}
\keyword{datasets}
