% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{donorsHBA}
\alias{donorsHBA}
\title{Donor ID's for Brains in the Human Brain Atlas}
\format{A character vector of length 6:}
\source{
\url{"http://human.brain-map.org/static/download"}
}
\usage{
donorsHBA
}
\description{
A vector of the donor ID's from the 6 brains in the human brain atlas
}
\keyword{datasets}
