% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genesHBA}
\alias{genesHBA}
\title{Genes Surveyed in the Human Brain Atlas}
\format{A large character vector of length 29131}
\source{
\url{"http://human.brain-map.org/static/download"}
}
\usage{
genesHBA
}
\description{
A list of the genes surved in the Human Brain Atlas at the Allen Institute
}
\keyword{datasets}
