% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ontologyHBA}
\alias{ontologyHBA}
\title{Ontology of the Human Brains in the Human Brain Atlas}
\format{A data frame with 559 rows and 2 variables:
\describe{
 \item{acronyms}{brain structure acronym as used in `subregionsHBA`}
 \item{name}{brain structure names; substructures may have same acronyms}
}}
\source{
\url{"http://human.brain-map.org/static/download"}
}
\usage{
ontologyHBA
}
\description{
A dataset containing the acronyms and corresponding structure names for all of the brain structures featured in the Human Brain Atlas.
}
\keyword{datasets}
