% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{probesHBA}
\alias{probesHBA}
\title{mRNA Probes Used in the Human Brain Atlas}
\format{A large character vector of length 58692}
\source{
\url{"http://human.brain-map.org/static/download"}
}
\usage{
probesHBA
}
\description{
A list of the mRNA probes surveyed in the Human Brain Atlas at the Allen Institute.
}
\keyword{datasets}
