% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subregionsHBA}
\alias{subregionsHBA}
\title{Brain Subregions Surveyed in the Human Brain Atlas}
\format{A character vector of length 232}
\source{
\url{"http://human.brain-map.org/static/download"}
}
\usage{
subregionsHBA
}
\description{
A list of the subregions acronyms surveyed in the Human Brain Atlas at the Allen Institute.
}
\keyword{datasets}
